Refactor #185 disable checkHost feature and move dtos
