+++ src/AppBundle/Model/User/DTO/LocaleSwitcherDTO.php
@@ -10,7 +10,7 @@
  * file that was distributed with this source code.
  */
 
-namespace AppBundle\Model\User\Locale\DTO;
+namespace AppBundle\Model\User\DTO;
 
 use AppBundle\Validator\Constraints as Assert;
