+++ src/AppBundle/Model/User/Registration/TwoStepRegistrationApproach.php
@@ -15,7 +15,7 @@ namespace AppBundle\Model\User\Registration;
 use AppBundle\Event\MailerEvent;
 use AppBundle\Exception\UserActivationException;
 use AppBundle\Model\User\Registration\Activation\ExpiredActivationProviderInterface;
-use AppBundle\Model\User\Registration\DTO\CreateUserDTO;
+use AppBundle\Model\User\DTO\CreateUserDTO;
 use AppBundle\Model\User\Registration\Generator\ActivationKeyCodeGeneratorInterface;
 use AppBundle\Model\User\Registration\NameSuggestion\Suggestor\SuggestorInterface;
 use AppBundle\Model\User\User;
@@ -188,7 +188,7 @@ final class TwoStepRegistrationApproach implements AccountCreationInterface, Acc
     /**
      * Builds a new user and validates it.
      *
-     * @param CreateUserDTO $userParameters
+     * @param \AppBundle\Model\User\DTO\CreateUserDTO $userParameters
      *
      * @return mixed[]
      */
